feat(poll) handle poll votes

+++ src/main/java/com/iseplife/api/services/PollService.java
@@ -1,9 +1,10 @@
 package com.iseplife.api.services;
 
-import com.iseplife.api.conf.jwt.TokenPayload;
 import com.iseplife.api.dao.poll.PollFactory;
 import com.iseplife.api.dto.embed.PollCreationDTO;
+import com.iseplife.api.dto.embed.view.PollChoiceView;
 import com.iseplife.api.dto.embed.view.PollView;
+import com.iseplife.api.entity.feed.Feed;
 import com.iseplife.api.entity.post.embed.poll.Poll;
 import com.iseplife.api.entity.post.embed.poll.PollChoice;
 import com.iseplife.api.entity.post.embed.poll.PollVote;
@@ -11,6 +12,7 @@ import com.iseplife.api.entity.user.Student;
 import com.iseplife.api.dao.poll.PollChoiceRepository;
 import com.iseplife.api.dao.poll.PollRepository;
 import com.iseplife.api.dao.poll.PollVoteRepository;
+import com.iseplife.api.exceptions.AuthException;
 import com.iseplife.api.exceptions.IllegalArgumentException;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,10 +24,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-/**
- * Created by Guillaume on 31/07/2017.
- * back
- */
 @Service
 public class PollService {
   @Autowired
@@ -44,43 +42,72 @@ public class PollService {
   @Autowired
   PostService postService;
 
-  public Poll getPoll(Long pollId) {
-    Optional<Poll> poll = pollRepository.findById(pollId);
-    if(poll.isEmpty())
-      throw new IllegalArgumentException("Could not find this poll (id:" + pollId + ")");
+
+  public Poll getPoll(Long id) {
+    Optional<Poll> poll = pollRepository.findById(id);
+    if (poll.isEmpty() || !SecurityService.hasReadAccess(poll.get()))
+      throw new IllegalArgumentException("Could not find this poll (id:" + id + ")");
 
     return poll.get();
   }
 
-  public void addVote(Long pollId, Long pollAnswId, Long studentId) {
+  public PollView getPollView(Long id){
+    Poll poll = getPoll(id);
+
+    return PollFactory.toView(poll);
+  }
+
+  public Poll bindPollToPost(Long id, Feed feed){
+    Poll poll = getPoll(id);
+    poll.setFeed(feed);
+
+    return pollRepository.save(poll);
+  }
+
+
+  public void addVote(Long pollId, Long choiceId, Long studentId) {
     Poll poll = getPoll(pollId);
 
+    if(poll.getEndsAt().before(new Date()))
+      throw new AuthException("You are not allowed to vote on this poll anymore");
+
+    if(pollVoteRepository.findByChoice_IdAndStudent_Id(choiceId, studentId).isPresent())
+      throw new IllegalArgumentException("This choice has already been chosen");
+
+
+    Optional<PollChoice> pollAnswer = pollChoiceRepository.findById(choiceId);
+    if (pollAnswer.isEmpty())
+      throw new IllegalArgumentException("Could not find this poll's choice (id:" + choiceId + ")");
+
     if (!poll.getMultiple()) {
-      List<PollVote> voteList = pollVoteRepository.findByAnswer_Poll_IdAndStudent_Id(pollId, studentId).stream()
-        .filter(votes -> !votes.getAnswer().getId().equals(pollAnswId))
+      List<PollVote> voteList = pollVoteRepository.findByChoice_Poll_IdAndStudent_Id(poll.getId(), studentId).stream()
+        .filter(votes -> !votes.getChoice().getId().equals(choiceId))
         .collect(Collectors.toList());
       pollVoteRepository.deleteAll(voteList);
     }
 
-    if (checkIsAnswered(pollAnswId, studentId)) {
-      throw new IllegalArgumentException("This answer has already been chosen");
-    }
-
-    Optional<PollChoice> pollAnswer = pollChoiceRepository.findById(pollAnswId);
-    if(pollAnswer.isEmpty())
-      throw new IllegalArgumentException("Unknown answer");
 
     Student student = studentService.getStudent(studentId);
     PollVote pollVote = new PollVote();
-    pollVote.setAnswer(pollAnswer.get());
+    pollVote.setChoice(pollAnswer.get());
     pollVote.setStudent(student);
     pollVoteRepository.save(pollVote);
   }
 
 
-  private boolean checkIsAnswered(Long answerId, Long userid) {
-    return pollVoteRepository.findByAnswer_IdAndStudent_Id(answerId, userid) != null;
+  public void removeVote(Long pollId, Long choiceId, Long studentId) {
+    Poll poll = getPoll(pollId);
+
+    if(poll.getEndsAt().before(new Date()))
+      throw new AuthException("You are not allowed to vote on this poll anymore");
+
+    Optional<PollVote> vote = pollVoteRepository.findByChoice_IdAndStudent_Id(choiceId, studentId);
+    if (vote.isEmpty())
+      throw new IllegalArgumentException("Could not find this answer's vote in this poll");
+
+    pollVoteRepository.delete(vote.get());
   }
+
 
   public PollView createPoll(PollCreationDTO dto) {
     ModelMapper mapper = new ModelMapper();
@@ -88,7 +115,9 @@ public class PollService {
 
     mapper.map(dto, poll);
 
+    poll.setCreation(new Date(
 
+    ));
     poll.setChoices(new ArrayList<>());
     dto.getChoices().forEach(q -> {
       PollChoice pollChoice = new PollChoice();
@@ -101,23 +130,12 @@ public class PollService {
   }
 
 
+  public List<PollChoiceView> getPollVotes(Long pollId) {
+    Poll poll = getPoll(pollId);
 
-  public List<PollVote> getUserVotes(Long pollId) {
-    return pollVoteRepository.findByAnswer_Poll_Id(pollId);
+    return poll.getChoices().stream()
+      .map(choice -> poll.getAnonymous() ? PollFactory.toAnonymousView(choice): PollFactory.toView(choice))
+      .collect(Collectors.toList());
   }
 
-  public List<PollVote> getVote(Long pollId, long studentId) {
-    return pollVoteRepository.findByAnswer_Poll_IdAndStudent_Id(pollId, studentId);
-  }
-
-  public void removeVote(Long id, Long answerId, TokenPayload auth) {
-    Poll poll = getPoll(id);
-    if (poll.getEndsAt().before(new Date())) {
-      throw new IllegalArgumentException("you cannot vote for this poll anymore");
-    }
-    PollVote vote = pollVoteRepository.findByAnswer_IdAndStudent_Id(answerId, auth.getId());
-    if (vote != null) {
-      pollVoteRepository.delete(vote);
-    }
-  }
 }
